+++ R/AggregateCovariates.R
@@ -18,9 +18,13 @@
 #'
 #' @param targetIds   A list of cohortIds for the target cohorts
 #' @param outcomeIds  A list of cohortIds for the outcome cohorts
-#' @param minPriorObservation The minimum time in the database a patient in the target cohorts must be observed prior to index
+#' @param minPriorObservation The minimum time (in days) in the database a patient in the target cohorts must be observed prior to index
+#' @param outcomeWashoutDays Patients with the outcome within outcomeWashout days prior to index are excluded from the risk factor analysis
 #' @template timeAtRisk
 #' @param covariateSettings   An object created using \code{FeatureExtraction::createCovariateSettings}
+#' @param caseCovariateSettings An object created using \code{createDuringCovariateSettings}
+#' @param casePreTargetDuration    The number of days prior to case index we use for FeatureExtraction
+#' @param casePostOutcomeDuration    The number of days prior to case index we use for FeatureExtraction
 #' @param minCharacterizationMean The minimum mean value for characterization output. Values below this will be cut off from output. This
 #'                                will help reduce the file size of the characterization output, but will remove information
 #'                                on covariates that have very low values. The default is 0.
@@ -33,12 +37,54 @@ createAggregateCovariateSettings <- function(
     targetIds,
     outcomeIds,
     minPriorObservation = 0,
+    outcomeWashoutDays = 0,
     riskWindowStart = 1,
     startAnchor = "cohort start",
     riskWindowEnd = 365,
     endAnchor = "cohort start",
-    covariateSettings,
-    minCharacterizationMean = 0) {
+    covariateSettings = FeatureExtraction::createCovariateSettings(
+      useDemographicsGender = T,
+      useDemographicsAge = T,
+      useDemographicsAgeGroup = T,
+      useDemographicsRace = T,
+      useDemographicsEthnicity = T,
+      useDemographicsIndexYear = T,
+      useDemographicsIndexMonth = T,
+      useDemographicsTimeInCohort = T,
+      useDemographicsPriorObservationTime = T,
+      useDemographicsPostObservationTime = T,
+      useConditionGroupEraLongTerm = T,
+      useDrugGroupEraOverlapping = T,
+      useDrugGroupEraLongTerm = T,
+      useProcedureOccurrenceLongTerm = T,
+      useMeasurementLongTerm = T,
+      useObservationLongTerm = T,
+      useDeviceExposureLongTerm = T,
+      useVisitConceptCountLongTerm = T,
+      useConditionGroupEraShortTerm = T,
+      useDrugGroupEraShortTerm = T,
+      useProcedureOccurrenceShortTerm = T,
+      useMeasurementShortTerm = T,
+      useObservationShortTerm = T,
+      useDeviceExposureShortTerm = T,
+      useVisitConceptCountShortTerm = T,
+      endDays = 0,
+      longTermStartDays =  -365,
+      shortTermStartDays = -30
+    ),
+    caseCovariateSettings = createDuringCovariateSettings(
+      useConditionGroupEraDuring = T,
+      useDrugGroupEraDuring = T,
+      useProcedureOccurrenceDuring = T,
+      useDeviceExposureDuring = T,
+      useMeasurementDuring = T,
+      useObservationDuring = T,
+      useVisitConceptCountDuring = T
+      ),
+    casePreTargetDuration = 365,
+    casePostOutcomeDuration = 365,
+    minCharacterizationMean = 0
+    ) {
   errorMessages <- checkmate::makeAssertCollection()
   # check targetIds is a vector of int/double
   .checkCohortIds(
@@ -53,14 +99,16 @@ createAggregateCovariateSettings <- function(
     errorMessages = errorMessages
   )
 
-  # check TAR
-  .checkTimeAtRisk(
-    riskWindowStart = riskWindowStart,
-    startAnchor = startAnchor,
-    riskWindowEnd = riskWindowEnd,
-    endAnchor = endAnchor,
-    errorMessages = errorMessages
-  )
+  # check TAR - EFF edit
+  for(i in 1:length(riskWindowStart)){
+    .checkTimeAtRisk(
+      riskWindowStart = riskWindowStart[i],
+      startAnchor = startAnchor[i],
+      riskWindowEnd = riskWindowEnd[i],
+      endAnchor = endAnchor[i],
+      errorMessages = errorMessages
+    )
+  }
 
   # check covariateSettings
   .checkCovariateSettings(
@@ -68,24 +116,49 @@ createAggregateCovariateSettings <- function(
     errorMessages = errorMessages
   )
 
+  # check temporal is false
+  if(inherits(covariateSettings, 'covariateSettings')){
+    covariateSettings <- list(covariateSettings)
+  }
+  if(sum(unlist(lapply(covariateSettings, function(x){x$temporal})))>0){
+    stop('Temporal covariateSettings not supported by createAggregateCovariateSettings()')
+  }
+
   # check minPriorObservation
   .checkMinPriorObservation(
     minPriorObservation = minPriorObservation,
     errorMessages = errorMessages
   )
 
+  # add check for outcomeWashoutDays
+
   checkmate::reportAssertions(errorMessages)
+
+  # check unique Ts and Os
+  if(length(targetIds) != length(unique(targetIds))){
+    message('targetIds have duplicates - making unique')
+    targetIds <- unique(targetIds)
+  }
+  if(length(outcomeIds) != length(unique(outcomeIds))){
+    message('outcomeIds have duplicates - making unique')
+    outcomeIds <- unique(outcomeIds)
+  }
+
 
   # create list
   result <- list(
     targetIds = targetIds,
     outcomeIds = outcomeIds,
     minPriorObservation = minPriorObservation,
+    outcomeWashoutDays = outcomeWashoutDays,
     riskWindowStart = riskWindowStart,
     startAnchor = startAnchor,
     riskWindowEnd = riskWindowEnd,
     endAnchor = endAnchor,
     covariateSettings = covariateSettings,
+    caseCovariateSettings = caseCovariateSettings,
+    casePreTargetDuration = casePreTargetDuration,
+    casePostOutcomeDuration = casePostOutcomeDuration,
     minCharacterizationMean = minCharacterizationMean
   )
 
@@ -122,6 +195,9 @@ computeAggregateCovariateAnalyses <- function(
     runId = 1) {
   # check inputs
 
+  # create a unique execution id using the runId - moved to runChar
+  #runId <- hashInt(settings = aggregateCovariateSettings)
+
   start <- Sys.time()
 
   connection <- DatabaseConnector::connect(
@@ -131,8 +207,49 @@ computeAggregateCovariateAnalyses <- function(
     DatabaseConnector::disconnect(connection)
   )
 
+
+  # creating cohort_details
+  message("Creating and inserting cohort details")
+  cohortDetails <- getCohortDetails(
+    targetIds = aggregateCovariateSettings$targetIds,
+    outcomeIds = aggregateCovariateSettings$outcomeIds
+    )
+
+  # for each time-at-risk get case details
+  for(i in 1:length(aggregateCovariateSettings$riskWindowStart)){
+    cohortDetails <- rbind(
+      cohortDetails,
+      getCaseDetails(
+        targetIds = aggregateCovariateSettings$targetIds,
+        outcomeIds = aggregateCovariateSettings$outcomeIds,
+        timeAtRiskId = i
+      ))
+  }
+
+  # using row_id for unique cohort id for FE
+  cohortDetails$cohortDefinitionId <- 1:nrow(cohortDetails)
+
+  # check cohortDefinitionIds are unique
+  if(sum(table(cohortDetails$cohortDefinitionIds)> 1) != 0){
+    stop('Unique constrain on cohortDefinitionIds failed')
+  }
+
+
+  DatabaseConnector::insertTable(
+    data = cohortDetails,
+    camelCaseToSnakeCase = T,
+    connection = connection,
+    tableName =  '#cohort_details',
+    tempTable = T,
+    dropTableIfExists = T,
+    createTable = T,
+    progressBar = T
+    )
+
   # select T, O, create TnO, TnOc, Onprior T
   # into temp table #agg_cohorts
+  message("Computing aggregate covariate cohorts")
+  start <- Sys.time()
   createCohortsOfInterest(
     connection = connection,
     dbms = connectionDetails$dbms,
@@ -144,9 +261,20 @@ computeAggregateCovariateAnalyses <- function(
     outcomeTable,
     tempEmulationSchema
   )
-
+  completionTime <- Sys.time() - start
+  message(paste0('Computing aggregate covariate cohorts took ',round(completionTime,digits = 1), ' ', units(completionTime)))
   ## get counts
-  sql <- "select cohort_definition_id, count(*) row_count, count(distinct subject_id) person_count from #agg_cohorts group by cohort_definition_id;"
+  message("Extracting cohort counts")
+  sql <- "select
+  cohort_definition_id,
+  count(*) row_count,
+  count(distinct subject_id) person_count,
+  min(datediff(day, cohort_start_date, cohort_end_date)) min_exposure_time,
+  avg(datediff(day, cohort_start_date, cohort_end_date)) mean_exposure_time,
+  max(datediff(day, cohort_start_date, cohort_end_date)) max_exposure_time
+  from
+  (select * from #agg_cohorts_before union select * from #agg_cohorts_extras) temp
+  group by cohort_definition_id;"
   sql <- SqlRender::translate(
     sql = sql,
     targetDialect = connectionDetails$dbms
@@ -157,20 +285,79 @@ computeAggregateCovariateAnalyses <- function(
     snakeCaseToCamelCase = T,
   )
 
-  message("Computing aggregate covariate results")
+  message("Computing aggregate before covariate results")
 
   result <- FeatureExtraction::getDbCovariateData(
     connection = connection,
     oracleTempSchema = tempEmulationSchema,
     cdmDatabaseSchema = cdmDatabaseSchema,
-    cohortTable = "#agg_cohorts",
+    cohortTable = "#agg_cohorts_before",
     cohortTableIsTemp = T,
-    cohortId = -1,
+    cohortIds = -1,
     covariateSettings = aggregateCovariateSettings$covariateSettings,
     cdmVersion = cdmVersion,
     aggregated = T,
     minCharacterizationMean = aggregateCovariateSettings$minCharacterizationMean
   )
+
+  message("Computing case covariate results")
+  resultCase <- FeatureExtraction::getDbCovariateData(
+    connection = connection,
+    oracleTempSchema = tempEmulationSchema,
+    cdmDatabaseSchema = cdmDatabaseSchema,
+    cohortTable = "#agg_cohorts_cases",
+    cohortTableIsTemp = T,
+    cohortIds = -1,
+    covariateSettings = aggregateCovariateSettings$caseCovariateSettings,
+    cdmVersion = cdmVersion,
+    aggregated = T,
+    minCharacterizationMean = aggregateCovariateSettings$minCharacterizationMean
+  )
+
+  # add covariates to table
+  if(!is.null(result$covariates)){
+    if (!is.null(resultCase$covariates)) {
+      Andromeda::appendToTable(
+        tbl = result$covariates,
+        data = resultCase$covariates
+      )
+    }
+  } else{
+    if (!is.null(resultCase$covariates)) {
+      result$covariates <- resultCase$covariates
+    }
+  }
+
+  # covariatesContinuous
+  if(!is.null(result$covariatesContinuous)){
+    if (!is.null(resultCase$covariatesContinuous)) {
+      Andromeda::appendToTable(
+        tbl = result$covariatesContinuous,
+        data = resultCase$covariatesContinuous
+      )
+    }
+  } else{
+    if (!is.null(resultCase$covariatesContinuous)) {
+      result$covariatesContinuous <- resultCase$covariatesContinuous
+    }
+  }
+
+  # update covariateRef
+  result$covariateRef <- unique(
+    rbind(
+      as.data.frame(result$covariateRef),
+      as.data.frame(resultCase$covariateRef)
+    )
+  )
+
+  # update analysisRef
+  result$analysisRef <- unique(
+    rbind(
+      as.data.frame(result$analysisRef),
+      as.data.frame(resultCase$analysisRef)
+    )
+  )
+
   # adding counts as a new table
   result$cohortCounts <- counts
 
@@ -219,13 +406,31 @@ computeAggregateCovariateAnalyses <- function(
     )
   )
 
+  caseCovariateSettingsJson <- as.character(
+    ParallelLogger::convertSettingsToJson(
+      aggregateCovariateSettings$caseCovariateSettings
+    )
+  )
+
   result$settings <- data.frame(
     runId = runId,
     databaseId = databaseId,
     covariateSettingJson = covariateSettingsJson,
+    caseCovariateSettingsJson = caseCovariateSettingsJson,
+    casePostOutcomeDuration = aggregateCovariateSettings$casePostOutcomeDuration,
+    casePreTargetDuration = aggregateCovariateSettings$casePreTargetDuration,
+    minPriorObservation = aggregateCovariateSettings$minPriorObservation,
+    outcomeWashoutDays = aggregateCovariateSettings$outcomeWashoutDays,
+    minCharacterizationMean = aggregateCovariateSettings$minCharacterizationMean
+  )
+
+  result$timeAtRisk <- data.frame(
+    runId = runId,
+    databaseId = databaseId,
+    timeAtRiskId = 1:length(aggregateCovariateSettings$riskWindowStart),
     riskWindowStart = aggregateCovariateSettings$riskWindowStart,
-    startAnchor = aggregateCovariateSettings$startAnchor,
     riskWindowEnd = aggregateCovariateSettings$riskWindowEnd,
+    startAnchor = aggregateCovariateSettings$startAnchor,
     endAnchor = aggregateCovariateSettings$endAnchor
   )
 
@@ -255,9 +460,12 @@ createCohortsOfInterest <- function(
     targetTable,
     outcomeDatabaseSchema,
     outcomeTable,
-    tempEmulationSchema) {
+    tempEmulationSchema
+    ) {
+
+  # first create Ts and Os calling RestrictCohorts.sql
   sql <- SqlRender::loadRenderTranslateSql(
-    sqlFilename = "createTargetOutcomeCombinations.sql",
+    sqlFilename = "RestrictCohorts.sql",
     packageName = "Characterization",
     dbms = dbms,
     cdm_database_schema = cdmDatabaseSchema,
@@ -269,18 +477,7 @@ createCohortsOfInterest <- function(
     target_ids = paste(aggregateCovariateSettings$targetIds, collapse = ",", sep = ","),
     outcome_ids = paste(aggregateCovariateSettings$outcomeIds, collapse = ",", sep = ","),
     min_prior_observation = aggregateCovariateSettings$minPriorObservation,
-    tar_start = aggregateCovariateSettings$riskWindowStart,
-    tar_start_anchor = ifelse(
-      aggregateCovariateSettings$startAnchor == "cohort start",
-      "cohort_start_date",
-      "cohort_end_date"
-    ),
-    tar_end = aggregateCovariateSettings$riskWindowEnd,
-    tar_end_anchor = ifelse(
-      aggregateCovariateSettings$endAnchor == "cohort start",
-      "cohort_start_date",
-      "cohort_end_date"
-    )
+    outcome_washout_days = aggregateCovariateSettings$outcomeWashoutDays
   )
 
   DatabaseConnector::executeSql(
@@ -289,4 +486,112 @@ createCohortsOfInterest <- function(
     progressBar = FALSE,
     reportOverallTime = FALSE
   )
+
+
+  # now create the cases per TAR
+  for(i in 1:length(aggregateCovariateSettings$riskWindowStart)){
+    sql <- SqlRender::loadRenderTranslateSql(
+      sqlFilename = "CreateCases.sql",
+      packageName = "Characterization",
+      dbms = dbms,
+      tempEmulationSchema = tempEmulationSchema,
+      tar_start = aggregateCovariateSettings$riskWindowStart[i],
+      tar_start_anchor = ifelse(
+        aggregateCovariateSettings$startAnchor[i] == "cohort start",
+        "cohort_start_date",
+        "cohort_end_date"
+      ),
+      tar_end = aggregateCovariateSettings$riskWindowEnd[i],
+      tar_end_anchor = ifelse(
+        aggregateCovariateSettings$endAnchor[i] == "cohort start",
+        "cohort_start_date",
+        "cohort_end_date"
+      ),
+      time_at_risk_id = i,
+      first = i==1,
+      case_post_outcome_duration = aggregateCovariateSettings$casePostOutcomeDuration,
+      case_pre_target_duration = aggregateCovariateSettings$casePreTargetDuration
+    )
+
+    DatabaseConnector::executeSql(
+      connection = connection,
+      sql = sql,
+      progressBar = FALSE,
+      reportOverallTime = FALSE
+    )
+  }
+
+}
+
+
+getCohortDetails <- function(
+    targetIds,
+    outcomeIds
+){
+  cohortDetails <- data.frame(
+    targetCohortId = c(rep(targetIds, 2), rep(0,2*length(outcomeIds))),
+    outcomeCohortId = c(rep(0,2*length(targetIds)), rep(outcomeIds, 2)),
+    cohortType = c(
+      rep('Tall', length(targetIds)),
+      rep('T', length(targetIds)),
+      rep('Oall', length(outcomeIds)),
+      rep('O', length(outcomeIds))
+    ),
+    timeAtRiskId = 0
+  )
+
+  comboTypes <- 'TnOprior'
+  for(comboType in comboTypes){
+    cohortDetailsExtra <- as.data.frame(
+      merge(
+        targetIds,
+        outcomeIds)
+    )
+    colnames(cohortDetailsExtra) <- c('targetCohortId', 'outcomeCohortId')
+    cohortDetailsExtra$cohortType <- comboType
+    cohortDetailsExtra$timeAtRiskId <- 0
+    cohortDetails <- rbind(cohortDetails, cohortDetailsExtra)
+  }
+
+  return(cohortDetails)
+}
+
+getCaseDetails <- function(
+    targetIds,
+    outcomeIds,
+    timeAtRiskId
+){
+
+  cohortDetails <- c()
+
+  comboTypes <- c('TnObetween','OnT', 'TnO')
+  for(comboType in comboTypes){
+    cohortDetailsExtra <- as.data.frame(
+      merge(
+        targetIds,
+        outcomeIds)
+    )
+    colnames(cohortDetailsExtra) <- c('targetCohortId', 'outcomeCohortId')
+    cohortDetailsExtra$cohortType <- comboType
+    cohortDetailsExtra$timeAtRiskId <- timeAtRiskId
+    cohortDetails <- rbind(cohortDetails, cohortDetailsExtra)
+  }
+
+  return(cohortDetails)
+}
+
+hex_to_int = function(h) {
+  xx = strsplit(tolower(h), "")[[1L]]
+  pos = match(xx, c(0L:9L, letters[1L:6L]))
+  sum((pos - 1L) * 16^(rev(seq_along(xx) - 1)))
+}
+
+hashInt <- function(settings){
+  a <- digest::digest(
+    object = settings,
+    algo='xxhash32',
+    seed=0
+  )
+  result <- hex_to_int(a)
+  return(result)
 }