+++ NEWS.md
@@ -1,3 +1,19 @@
+Characterization 0.3.1
+======================
+- Removed DatabaseConnector from Remotes in DESCRIPTION. Fixes GitHb issue 38.
+- Added check to covariateSettings input in createAggregateCovariateSettings to error if temporal is T
+- adding during cohort covariate settings
+- added a case covariate settings inputs to aggregate covariates
+- added casePreTargetDuration and casePostTreatmentDuration integer inputs to aggregate covariates
+
+Characterization 0.3.0
+======================
+- Added new outcomeWashoutDays parameter to createAggregateCovariateSettings to remove outcome occurances that are continuations of a prior outcome occurrence
+- Changed the way cohort definition ids are created in createAggregateCovariateSettings to use hash of target id, outcome id and type.  This lets users combine different studies into a single result database.
+- Added database migration capability and created new migrations for the recent updates.
+
+
+
 Characterization 0.2.0
 ======================
 Updated dependency to FeatureExtraction (>= 3.5.0) to support minCharacterizationMean paramater.